Inventaire sans ScriptableObject OK

+++ Assets/Scripts/Controllers/BattleControllerData.cs
@@ -3,7 +3,7 @@
     [System.Serializable]
     public class BattleControllerData
     {
-        public int currentFrame;
+        public ulong currentFrame;
         public BattleState battleState;
 
         public BattleControllerData(BattleController battleController)
